adding more tests and refactoring
