+++ src/MLOps.NET/Storage/Deployments/Interfaces/IModelPathGenerator.cs
