+++ test/MLOps.NET.Tests/MLOps.NET.Tests.csproj
@@ -18,6 +18,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\MLOps.NET\MLOps.NET.csproj" />
+    <ProjectReference Include="..\MLOps.NET.Tests.Common\MLOps.NET.Tests.Common.csproj" />
   </ItemGroup>
 
 </Project>
